Conflicts:
	examples/directx11_example/imgui_impl_dx11.cpp
	examples/directx9_example/imgui_impl_dx9.cpp
	examples/opengl3_example/imgui_impl_glfw_gl3.cpp
	examples/opengl_example/imgui_impl_glfw.cpp
	imgui.cpp

+++ examples/directx11_example/imgui_impl_dx11.cpp
@@ -119,7 +119,7 @@ static void ImGui_ImplDX11_RenderDrawLists(ImDrawData* draw_data)
     for (int n = 0; n < draw_data->cmd_lists_count; n++)
     {
         const ImDrawList* cmd_list = draw_data->cmd_lists[n];
-        for (size_t cmd_i = 0; cmd_i < cmd_list->cmd_buffer.size(); cmd_i++)
+        for (int cmd_i = 0; cmd_i < cmd_list->cmd_buffer.size(); cmd_i++)
         {
             const ImDrawCmd* pcmd = &cmd_list->cmd_buffer[cmd_i];
             if (pcmd->user_callback)
